// subgraph/city/generated/CityHistory/CityHistory.ts
export class AetherUseRecorded extends ethereum.Event {
  get params(): AetherUseRecorded__Params {
    return new AetherUseRecorded__Params(this);
  }
}

export class AetherUseRecorded__Params {
  _event: AetherUseRecorded;

  constructor(event: AetherUseRecorded) {
    this._event = event;
  }

  get plotId(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get totalAetherUses(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class LayerAdded extends ethereum.Event {
  get params(): LayerAdded__Params {
    return new LayerAdded__Params(this);
  }
}

export class LayerAdded__Params {
  _event: LayerAdded;

  constructor(event: LayerAdded) {
    this._event = event;
  }

  get plotId(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get newLayerCount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class OwnershipTransferRecorded extends ethereum.Event {
  get params(): OwnershipTransferRecorded__Params {
    return new OwnershipTransferRecorded__Params(this);
  }
}

export class OwnershipTransferRecorded__Params {
  _event: OwnershipTransferRecorded;

  constructor(event: OwnershipTransferRecorded) {
    this._event = event;
  }

  get plotId(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get transferCount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class OwnershipTransferred extends ethereum.Event {
  get params(): OwnershipTransferred__Params {
    return new OwnershipTransferred__Params(this);
  }
}

export class OwnershipTransferred__Params {
  _event: OwnershipTransferred;

  constructor(event: OwnershipTransferred) {
    this._event = event;
  }

  get previousOwner(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get newOwner(): Address {
    return this._event.parameters[1].value.toAddress();
  }
}

export class PlotHistoryInitialized extends ethereum.Event {
  get params(): PlotHistoryInitialized__Params {
    return new PlotHistoryInitialized__Params(this);
  }
}

export class PlotHistoryInitialized__Params {
  _event: PlotHistoryInitialized;

  constructor(event: PlotHistoryInitialized) {
    this._event = event;
  }

  get plotId(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get firstBuilder(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get faction(): i32 {
    return this._event.parameters[2].value.toI32();
  }

  get genesisEra(): boolean {
    return this._event.parameters[3].value.toBoolean();
  }
}

export class CityHistory__provenanceOfResult {
  value0: Address;
  value1: BigInt;
  value2: BigInt;
  value3: BigInt;
  value4: BigInt;
  value5: BigInt;
  value6: i32;
  value7: boolean;

  constructor(
    value0: Address,
    value1: BigInt,
    value2: BigInt,
    value3: BigInt,
    value4: BigInt,
    value5: BigInt,
    value6: i32,
    value7: boolean,
  ) {
    this.value0 = value0;
    this.value1 = value1;
    this.value2 = value2;
    this.value3 = value3;
    this.value4 = value4;
    this.value5 = value5;
    this.value6 = value6;
    this.value7 = value7;
  }

  toMap(): TypedMap<string, ethereum.Value> {
    let map = new TypedMap<string, ethereum.Value>();
    map.set("value0", ethereum.Value.fromAddress(this.value0));
    map.set("value1", ethereum.Value.fromUnsignedBigInt(this.value1));
    map.set("value2", ethereum.Value.fromUnsignedBigInt(this.value2));
    map.set("value3", ethereum.Value.fromUnsignedBigInt(this.value3));
    map.set("value4", ethereum.Value.fromUnsignedBigInt(this.value4));
    map.set("value5", ethereum.Value.fromUnsignedBigInt(this.value5));
    map.set(
      "value6",
      ethereum.Value.fromUnsignedBigInt(BigInt.fromI32(this.value6)),
    );
    map.set("value7", ethereum.Value.fromBoolean(this.value7));
    return map;
  }

  getFirstBuilder(): Address {
    return this.value0;
  }

  getCreatedAt(): BigInt {
    return this.value1;
  }

  getLayerCount(): BigInt {
    return this.value2;
  }

  getOwnershipTransfers(): BigInt {
    return this.value3;
  }

  getAetherUses(): BigInt {
    return this.value4;
  }

  getHistoricScore(): BigInt {
    return this.value5;
  }

  getOriginFaction(): i32 {
    return this.value6;
  }

  getGenesisEra(): boolean {
    return this.value7;
  }
}

export class CityHistory extends ethereum.SmartContract {
  static bind(address: Address): CityHistory {
    return new CityHistory("CityHistory", address);
  }

  authorizedCallers(param0: Address): boolean {
    let result = super.call(
      "authorizedCallers",
      "authorizedCallers(address):(bool)",
      [ethereum.Value.fromAddress(param0)],
    );

    return result[0].toBoolean();
  }

  try_authorizedCallers(param0: Address): ethereum.CallResult<boolean> {
    let result = super.tryCall(
      "authorizedCallers",
      "authorizedCallers(address):(bool)",
      [ethereum.Value.fromAddress(param0)],
    );
    if (result.reverted) {
      return new ethereum.CallResult();
    }
    let value = result.value;
    return ethereum.CallResult.fromValue(value[0].toBoolean());
  }

  cityRegistry(): Address {
    let result = super.call("cityRegistry", "cityRegistry():(address)", []);

    return result[0].toAddress();
  }

  try_cityRegistry(): ethereum.CallResult<Address> {
    let result = super.tryCall("cityRegistry", "cityRegistry():(address)", []);
    if (result.reverted) {
      return new ethereum.CallResult();
    }
    let value = result.value;
    return ethereum.CallResult.fromValue(value[0].toAddress());
  }

  owner(): Address {
    let result = super.call("owner", "owner():(address)", []);

    return result[0].toAddress();
  }

  try_owner(): ethereum.CallResult<Address> {
    let result = super.tryCall("owner", "owner():(address)", []);
    if (result.reverted) {
      return new ethereum.CallResult();
    }
    let value = result.value;
    return ethereum.CallResult.fromValue(value[0].toAddress());
  }

  provenanceOf(param0: BigInt): CityHistory__provenanceOfResult {
    let result = super.call(
      "provenanceOf",
      "provenanceOf(uint256):(address,uint64,uint32,uint32,uint32,uint32,uint8,bool)",
      [ethereum.Value.fromUnsignedBigInt(param0)],
    );

    return new CityHistory__provenanceOfResult(
      result[0].toAddress(),
      result[1].toBigInt(),
      result[2].toBigInt(),
      result[3].toBigInt(),
      result[4].toBigInt(),
      result[5].toBigInt(),
      result[6].toI32(),
      result[7].toBoolean(),
    );
  }

  try_provenanceOf(
    param0: BigInt,
  ): ethereum.CallResult<CityHistory__provenanceOfResult> {
    let result = super.tryCall(
      "provenanceOf",
      "provenanceOf(uint256):(address,uint64,uint32,uint32,uint32,uint32,uint8,bool)",
      [ethereum.Value.fromUnsignedBigInt(param0)],
    );
    if (result.reverted) {
      return new ethereum.CallResult();
    }
    let value = result.value;
    return ethereum.CallResult.fromValue(
      new CityHistory__provenanceOfResult(
        value[0].toAddress(),
        value[1].toBigInt(),
        value[2].toBigInt(),
        value[3].toBigInt(),
        value[4].toBigInt(),
        value[5].toBigInt(),
        value[6].toI32(),
        value[7].toBoolean(),
      ),
    );
  }
}